requirements_dataset.txt
tensorflow>=2.10.0
opencv-python>=4.6.0
Pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.4.0
tqdm>=4.64.0
urllib3>=1.26.0
lxml>=4.9.0 